requirements.txt
pillow==7.0.0
requests==2.23.0
wget==3.2
progressbar2==3.47.0
